Moved member list to InGroupSection
