+++ calendar-app/Screens/GroupScreen.jsx
@@ -69,15 +69,6 @@ const GroupScreen = () => {
     });
   };
 
-  const renderMember = ({ item }) => (
-    <View style={styles.memberItem}>
-      <Text style={styles.memberText}>{item.username}</Text>
-      <TouchableOpacity onPress={() => alert('Remove member')}>
-        <Text style={styles.removeIcon}>×</Text>
-      </TouchableOpacity>
-    </View>
-  );
-
   return (
     <View style={styles.view}>
       <Header title={"Group"} />
@@ -88,15 +79,7 @@ const GroupScreen = () => {
           isLoading ? (
             <ActivityIndicator size="large" color={GlobalSecondaryColor}  style={styles.topContainer}/> 
           ) : inGroup ? (
-          <View style={styles.container}>
-            <Text style={styles.header}>Members</Text>
-            <FlatList
-              data={members}
-              renderItem={renderMember}
-              keyExtractor={item => item.id}
-              style={styles.membersList}
-            />
-          </View>
+            <InGroupSection />
           ) : (
             <NoGroupSection setGroupName={setGroupName} handleCreateGroup={handleCreateGroup} />
           )
@@ -121,41 +104,6 @@ const styles = StyleSheet.create({
     padding: 20,
     gap: 20,
     alignSelf: 'center',
-  },
-  container: {
-    backgroundColor: '#8EBBFF', 
-    borderRadius: 10,
-    padding: 10,
-    margin: 20,
-    alignItems: 'stretch',
-  },
-  header: {
-    color: GlobalTextColor,
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 10,
-    textAlign: 'center',
-  },
-  membersList: {
-    flexGrow: 0, 
-  },
-  memberItem: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: 10,
-    marginVertical: 5,
-    backgroundColor: '#8EBBFF',
-  },
-  memberText: {
-    color: GlobalTextColor, 
-    fontSize: 20,
-  },
-  removeIcon: {
-    color: '#FFFFFF', 
-    fontSize: 24,
-    fontWeight: 'bold',
-    padding: 8,
   },
 });
